fix(core): use integer arithmetic for payment money math
Replaces float arithmetic in application-fee calculation, refund-amount
parsing, and email receipt formatting with integer-only math, eliminating
IEEE-754 rounding hazards. Widens booking-payment and meeting-type money
columns to bigint so high-value cumulative totals cannot overflow.
Refund amount input now rejects scientific notation and sub-cent precision.

+++ priv/repo/migrations/20260508131127_add_payment_fields_to_meeting_types.exs
@@ -9,7 +9,7 @@ defmodule Tymeslot.Repo.Migrations.AddPaymentFieldsToMeetingTypes do
   def change do
     alter table(:meeting_types) do
       add(:payment_required, :boolean, default: false, null: false)
-      add(:price_cents, :integer)
+      add(:price_cents, :bigint)
       add(:is_archived, :boolean, default: false, null: false)
     end
